+++ lib/galaxy/util/__init__.py
@@ -38,8 +38,6 @@ import docutils.writers.html4css1
 
 from xml.etree import ElementTree, ElementInclude
 
-import wchartype
-
 from .inflection import Inflector, English
 inflector = Inflector(English)
 
@@ -56,22 +54,6 @@ bz2_magic = 'BZh'
 DEFAULT_ENCODING = os.environ.get('GALAXY_DEFAULT_ENCODING', 'utf-8')
 NULL_CHAR = '\000'
 BINARY_CHARS = [ NULL_CHAR ]
-
-
-def is_multi_byte( chars ):
-    for char in chars:
-        try:
-            char = unicode( char )
-        except UnicodeDecodeError:
-            # Probably binary
-            return False
-        if ( wchartype.is_asian( char ) or wchartype.is_full_width( char ) or
-                wchartype.is_kanji( char ) or wchartype.is_hiragana( char ) or
-                wchartype.is_katakana( char ) or wchartype.is_half_katakana( char ) or
-                wchartype.is_hangul( char ) or wchartype.is_full_digit( char ) or
-                wchartype.is_full_letter( char )):
-            return True
-    return False
 
 
 def is_binary( value, binary_chars=None ):
