+++ tools/data_source/upload.py
@@ -22,6 +22,7 @@ from galaxy.datatypes.checkers import check_binary, check_bz2, check_gzip, check
 from galaxy.datatypes.registry import Registry
 from galaxy.datatypes.util.image_util import get_image_ext
 from galaxy.util.json import dumps, loads
+from galaxy.util import multi_byte
 
 try:
     import Image as PIL
@@ -111,7 +112,7 @@ def add_file( dataset, registry, json_file, output_path ):
     if not dataset.type == 'url':
         # Already set is_multi_byte above if type == 'url'
         try:
-            dataset.is_multi_byte = util.is_multi_byte( codecs.open( dataset.path, 'r', 'utf-8' ).read( 100 ) )
+            dataset.is_multi_byte = multi_byte.is_multi_byte( codecs.open( dataset.path, 'r', 'utf-8' ).read( 100 ) )
         except UnicodeDecodeError, e:
             dataset.is_multi_byte = False
     # Is dataset an image?
